refactor(mobile): split FlappingBee into its own file
Addresses Codex review on #5059. Two structural fixes, no behavior change:

- Move FlappingBee and its private painter out of tappable_flapping_bee.dart
  into shared/widgets/flapping_bee.dart, restoring one public widget per file.
- Document the public properties on FlappingBee, and BeeRefreshIndicator's
  onRefresh and child.
